chore: restructure folders and files
