+++ packages/blade/src/storybook-recipes/AccessibilityLayer/NativeAria.native.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Pressable, Text, View } from 'react-native';
-import mapA11yProps from '../../utils/mapProps';
+import mapA11yProps from '../../utils/mapA11yProps';
 import { CloseIcon, EyeIcon } from '../../components/Icons';
 import { getCheckboxA11yProps } from './getCheckboxA11yProps';
