+++ packages/blade/src/utils/mapA11yProps/index.ts
@@ -1,0 +1,1 @@
+export { default } from './mapA11yProps';
