--- packages/blade/src/utils/mapProps/index.ts
@@ -1,1 +1,0 @@
-export { default } from './mapProps';
