--- packages/blade/src/utils/mapProps/mapProps.d.ts
@@ -1,1 +1,0 @@
-export { default } from './mapProps.web';
